refactor(git-summary): split modules and add tests
- Split git-summary into app/cli/dates/git/summary modules.
- Add unit tests for date ranges and summary parsing.
