+++ crates/git-summary/Cargo.toml
@@ -11,4 +11,5 @@ nils-term = { path = "../nils-term" }
 
 [dev-dependencies]
 nils-test-support = { path = "../nils-test-support" }
+pretty_assertions = { workspace = true }
 tempfile = "3"
