+++ crates/git-summary/src/git.rs
@@ -1,0 +1,42 @@
+use anyhow::{Context, Result};
+use std::process::{Command, Stdio};
+
+pub fn require_git() -> Result<(), &'static str> {
+    if Command::new("git")
+        .arg("--version")
+        .stdout(Stdio::null())
+        .stderr(Stdio::null())
+        .status()
+        .is_err()
+    {
+        return Err("❗ git is required but was not found in PATH.");
+    }
+
+    let status = Command::new("git")
+        .args(["rev-parse", "--git-dir"])
+        .stdout(Stdio::null())
+        .stderr(Stdio::null())
+        .status();
+
+    if !status.map(|s| s.success()).unwrap_or(false) {
+        return Err("⚠️ Not a Git repository. Run this command inside a Git project.");
+    }
+
+    Ok(())
+}
+
+pub fn run_git(args: &[String]) -> Result<String> {
+    let output = Command::new("git")
+        .args(args)
+        .stdout(Stdio::piped())
+        .stderr(Stdio::piped())
+        .output()
+        .with_context(|| format!("git {args:?}"))?;
+
+    if !output.status.success() {
+        let stderr = String::from_utf8_lossy(&output.stderr);
+        anyhow::bail!("git {args:?} failed: {stderr}");
+    }
+
+    Ok(String::from_utf8_lossy(&output.stdout).to_string())
+}
